registry: dispatch on `(service, procedure)` tuple (#101)
This namespaces procedure registration and dispatch to the servcie name. By
default, the name specified when building the RPC is used. This makes it
possible to support more complex forwarding use cases in the future.

The interface for `{raw, json, thrift}.Register` does not change here. We can
add service name customization in the future if necessary.

Depends on #100