Client side errors and crossdock tests (#114)
This implements support for consistent errors on the client side. Clients will
only see `transport.UnexpectedError` or `transport.BadRequestError`.

The design for errors is:

- Certain error types are `HandlerError`s. The Inbound has to know how to
  handle these. Examples: BadRequest and Unexpected.
- Other error types know how to convert themselves into HandlerErrors by
  implementing proving a `AsHandlerError() error` function. For example,
  UnknownProcedureError knows that it should be converted into a
  BadRequestError.
- All other error types are converted into UnexpectedErrors with information
  about the procedure and service.

The hierarchy we have is,

```
BadRequestError                 HandlerError
 |                                        |
 +--------> localBadRequestError <--------+
 |                                        |
 +--------> remoteBadRequestError         |
                                          |
UnexpectedError                           |
 |                                        |
 +--------> localUnexpectedError <--------+
 |
 +--------> remoteUnexpectedError
```

Note that only the `local` versions of the error types are valid
`HandlerError`s. That is, only the local versions make it to the Inbound as-is
and must be handled (resulting in a 400-level error for `BadRequest` and
500-level for `Unexpected`). If a remote version of either error is returned by
a `Handler`, it causes a *new* `UnexpectedError` which includes the procedure
name and service for that call. This way, we end up with full information for
multi-hop failures:

    UnexpectedError: error for procedure "foo" of service "a":
        UnexpectedError: error for procedure "bar" of service "b":
            BadRequest: unrecognized procedure "baz" for service "c"