headers: Verify headers exactly in all tests
Previously, we were checking that the headers in the request were a superset
of what we expected because we were not namespacing application headers to
`Rpc-Header-*` in HTTP, so we had to allow for extra headers like
`User-Agent`, `Content-Length`, etc.

Now that we're using namespaced HTTP headers, we can check that application
headers match *exactly* what we expect.