Hide error constructors
This hides constructors for transport errors in an internal package, reducing
the error-related surface area of the transport package to:

- `transport.IsBadRequestError`
- `transport.IsUnexpectedError`
- `transport.IsTimeoutError`

Note that previously, the way to check if something was one of these errors
was to type-cast into that interface. Since there's no behavior attached to
these errors at this time, I think it's safe to switch to the `Is*` approach.

Resolves https://github.com/yarpc/yarpc-go/issues/112