Gnarly hack for invalid stack pointer error (#218)
In short: A bug in Go 1.5 causes a specific form of comparison
(runtime.assertE2T2) to leave an invalid pointer on the stack instead
of zeroing it out. Usually, this isn't a problem because the function
returns afterwards. However, we're consistently hitting a case where
another function call is causing the stack to be grown while the
pointer is still invalid. The scanner responsible for copying the
stack as part of growing it runs into the invalid pointer and
crashes.

To work around this, we grow the stack significantly beforehand to
reduce the likelihood of another growth attempt while the pointer is
invalid.