transport: Add Options for transport-based customization
Transport options are intended to provide an extension point for outbounds and
inbounds so that they can customize the behavior of another part of the system
in a very decoupled fashion. Only the transport that wants to specialize
behavior needs to know about the specialization.

The use case that inspired this: We want to disable Thrift enveloping only for
TChannel. With this change, TChannel will be able to set a property on Options
that informs the Thrift encoding that these requests should not be enveloped.

Options are currently provided from the following sources:

-   Outbounds have Options scoped to all requests send through that outbound
    and their responses.
-   ResponseWriters have Options scoped to a specific request and its
    response.