transport: Add Options for transport-based customization (#255)
Transport options are intended to provide an extension point for outbounds and
inbounds so that they can customize the behavior of another part of the system
in a very decoupled fashion. Only the transport that wants to specialize
behavior, and the component that acts differently based on that option, need
to know about the specialization. All other parts of the system can simply
return zero-value Options.

The use case that inspired this: We want to disable Thrift enveloping only for
TChannel. With this change, TChannel will be able to set a property on Options
that informs the Thrift encoding that these requests should not be enveloped.

Options are currently provided from the following sources:

-   Outbounds have an `Options()` method which specifies options that apply to
    all requests sent through that outbound. This makes stateful outbounds
    difficult (see #260).
-   Inbounds pass in options that apply to a specific request down to the
    Handler handling that request.

See #256 for an example of how this will be used.