Handling of remote timeout
Before this commit, a remote handler that timeouts would return an
unexpected error.

An example when this is a problem:
 - A client starts a request with a timeout of X
 - The remote receives it and runs the remote handler
 - The handler reaches the deadline, and gracefully return a deadline
   error.
 - There is two possible path on the client:
   - The client timeouts itself slightly before receiving the response
     from the server. The user code receives a yarpc timeout error.
   - The client receives the deadline error from the server. The user
     code should receive a timeout error, but was receiving an
     unexpected remote error.

On the HTTP transport, the server returns a 504 with the message "remote
timeout". The client can then build a timeout error using the remote
message as the error message.

On the tchannel transport, the server returns a tchannel timeout error,
with the same message "remote timeout". The timeout error is built like
above.

It is interesting to note that a local timeout (ie, one generated by
tchannel or HTTP on the client side) will have the message "timeout".

Finally, any proxy along the way could also return a timeout error,
likely with a different message.

Programmatically the user code on the client always receives a
TimeoutError, and the error message can be used by humans for analysis.