[GRPC][Part 2] Create passthrough codec for GRPC
Summary: By default, GRPC requires a codec to convert requests over the
wire into protobufs/json/thrift/etc.  Since we want to use the YARPC
encoders to perform the conversions, we need a passthrough Codec to pass
to GRPC, so GRPC can perform the transport, and we'll pick the encoding
based on the YARPC configuration.

*Note* The "String" function is applied in gRPC as a content header e.g.
"-protobuf" or "-raw".  For now we put everything into "raw" but
conceivably in the future we'll want to customize the Codec "name" we
use to have compatibility with regular GRPC servers/clients

Test Plan: Wrote tests