[GRPC][Part 7] Add Caller and Encoding Headers to GRPC calls
Summary: Now that we have hardcoded raw functionality for yarpc over
gRPC, this diff add the ability to pass "Caller" and "Encoding" Headers
over the gRPC request.  In an ideal world, we would be able to set the
encoding directly in gRPC via a codec, but unfortunately it doesn't look
possible at this moment (maybe with a PR to gRPC).  The intermediary
state is a combination of the passthrough codec and adding this encoding
header so we can know what the encoding of the body is.

Test Plan: Added crossdock tests for json,thrift, and thriftgauntlet
proving we can handle the multiple encodings.  Also added a standalone
thrift example and ran the yclient/yserver to ensure the encodings &
callers were being pass through