[Opentracing Baggage][part 1] Use Opentracing for ctxpropagation crossdock test
Summary: We are moving away from using yarpc specific baggage for RPC
calls.  Instead the defacto way to use baggage in RPC will be to assign
baggage to the context through opentracing.  This PR changes our
ctxpropagation test to use the Opentracing api for getting/setting
baggage.

Test Plan: Tests Run, not sure if the full crossdock will work, eager to
see on the remote branch