inbound: Start() should accept a ServiceDetail (#383)
This changes `Inbound.Start` to accept a `ServiceDetail` rather than a
`Handler`. The `ServiceDetail` contains the name of the service and the
registry of procedures.

The motivation for this change is to give inbounds more information about the
methods they are handling. This is needed for oneway (#327, #378) because the
handler types are different and the inbound needs to know that in advance
based on the procedure name.

The service name is attached to the `ServiceDetail` so that the TChannel
transport can instantiate its own Channel with that service name if necessary.

This should also allow @willhug to drop the type casting of the `Handler` to
`Registry` in #354 and just use the `Registry` directly.