[GRPC][Part 4.5] Fix Lint after golint upgrade
Summary: In a recent commit to golint, there was a new rule added that
asserts that context.Context must be the first parameter in a function
definition.  This breaks some of GRPC's functionality since the handler
interface they require has ctx in the second position.  As a result, we
are fixing the cases where we have ctx in the second position, and we're
isolating the GRPC function we provide into it's own file and ignoring
it in our linting