Allow middleware and wrappers to cancel calls
`OutboundCall.BuildRequest` makes it impossible for middleware or outbound
wrappers higher up the stack to not call the underlying outbound because they
need to do that to get the Options, but by that point the underlying outbound
has control of the program flow and there's no way to cancel that request.
This adds another callback that the `BuildRequest` (now called `WithRequest`)
should call in the form of the `RequestSender` object.