transport: Remove Options completely
This removes `transport.Options` and simplifies the `transport.Outbound`
and `transport.Handler` interfaces. The only reason we even had
`transport.Options` was to encode Thrift bodies differently based on
whether we were on HTTP or TChannel. After significant discussion, we've
decided that the complexity caused by this is not worth the feature it
supports. Users can simply explicitly enable/disable enveloping for
compatibility with Apache Thrift.