[PeerList][Part 2] Add HostPortPeer implementations
Summary: This PR creates a concrete implementation of the Peer interface
from the previous PR.  The HostPortPeer is meant to be a generic peer
object, it has a string identifier for a hostport and can have multiple
subscriptions on top of it.
Since the retain/releases need to be controlled at the Agent layer we
don't have a mutex around changes to the References (Only the agent
should have access to doing that, and the agent needs the mutex since it
will decide whether to delete the peer altogether which the peer can't
do)
The next diff will implement the http agent that can be used to create
these peers.