[PeerList][Part 3] Adding a http PeerAgent
Summary: This diff introduces the concept of an http "PeerAgent".
Instances of this agent have the ability to create hostportpeers for
peerlists so they can be used in downstream calls. It controls the whole
lifecycle of a peer, as PeerLists should call retain/release during the
lifecycle of their peers.