[PeerList][Part 6] Shim old http outbounds with PeerLists
Summary: This diff converts the old HTTP.outbound to support PeerLists
and PeerAgents under the hood.  The API has not changed at all, but that
will happen in another diff. For now, this adds the support for the
HTTPOutbound to call into the PeerList to select the next peer for a
request