[PeerList][Part 7] Add round robin peer list
Summary: This diff introduces a basic version of a round robin peer
list. It uses a circular doubly linked list of nodes with a map from
peer identifier to the node holding that peer. This way we can rotate
through the peer list is a reproducible way.  There are maaaaany
optimizations we can do to this implementation, but this is meant to be
a proof of concept for the first version of a "true" PeerList.
The next PR will introduce the Add/Remove interface of a MutablePeerList