[PeerLists][Part 1] Add interfaces for peers/peerlists/peeragents (#399)
* [PeerLists][Part 1] Add interfaces for peers/peerlists/peeragents

Summary: This diff adds the first version of the interfaces of the
Peers, PeerLists, and PeerAgents.  Ideally the comments are
self-explanatory (please comment if they are not).  But essentially, a
PeerAgent controls the lifecycle of a Peer from creation to deletion.
PeerSubscribers request a peer from a PeerAgent using a
PeerIdentifier (Should be a way to uniquely identify a peer, e.g.
hostport) When the subscriber requests a new peer, they pass themselves
into the PeerAgent so that the agent can tell the Peer that someone is
subscribing to it.  Once all subscriptions to a peer are gone, it is
safe to delete that peer.  This is all the interface level, I'll add
implementation details in the next PRs.

*Note* this does not include the API for adding/removing from a
PeerList, that will be in a subsequent diff.