[PeerList][Part 11] Differentiating between Available/Unavailable Peers in RoundRobin
Summary: In order to have a more powerful PeerList in the round robin,
we need to be able to distinguish between a Peer that cannot accept
requests and one that can.  This PR adds a distinction in the RoundRobin
PeerList that will only return Available Peers from the ChoosePeer
method.

In the process of doing this, I realized it would be a bad idea to trust
the Peer we pass in the NotifyStatusChanged function, so I changed it
from a Peer to a PeerIdentifier so that the lists need to look up their
own object so we don't get an issue of duplicate peers getting passed
around with different state.

Test Plan: I've upgraded the PeerListAction tests to have a
NotifyStatusChangeAction which will update a Peer's Status and run the
NotifyStatusChange function on the PeerList with the provided Peer.
This required weaving a list of available MockPeers in the Apply
function.
Additionally, the gomock MockPeer was not very flexible with changing
the returned result from an Expectation, so I ended up creating a thin
mock of Peer instead.