[PeerList][Part 12] Move everything to yarpc/peer
Summary: In our original design we were putting all the peer logic into
the transports folder, but, since peer logic is orthogonal to the
specific transport we thought it was a good idea to split out all of
this logic into it's own subfolder.

I diverged a bit from the design in #444 because it had issues with
circular imports between the peertest/peer repos (similar issue with
hostport/lists).  Those ended up going into subfolders which I think
should be fine.

Additionally, some errors that were internal before are not external
because they are used by the http.Agent or http.Outbound

Test Plan: All the tests pass wheeeeee~

Task: #444