[PeerList][Part 7] Add round robin peer list (#405)
Summary: This diff introduces a round robin peer
list. It uses the container/ring package with a map from
peer identifier to the node holding that peer. This way we can rotate
through the peer list is a reproducible way. There are maaaaany
optimizations we can do to this implementation, but this is meant to be
a proof of concept for the first version of a "true" PeerList.
The next PR will introduce a blocking channel on ChoosePeer so that we can block requests until a peer is available (or there is a timeout)