[PeerList][Part 9] Move roundrobin mutex locks from peerring to public roundrobin funcs (#442)
Summary: In the initial revision the locks for the RoundRobin list were
only around the ring (which was fine). But with some changes that are
coming (moving peers in lists based on the peer's connection status) we
need to have locks at the top level, so we can move the peerring locks
up one level.