[PeerList][Part 10] Move PeerSubscriber management functions into Peer (#443)
Summary: Due to a potential deadlock situation when a notification
action occurs in the Agent Lock at the same time an Add/Remove occurs in
the PeerList we needed to move the location of PeerSubscription
management into the Peer level. This means that Notifications for peers
will not be entirely threadsafe, but that should be ok. Since the only
actions that should affect the Peer's subscribers(lists) are Retain/Releases
(which should be triggered by the lists) most inconsistency situations
can safely be ignored (event after a release) or won't happen fast
enough to matter (If an action occurs on a peer while you are retaining
it, it won't matter because you're going to check it's state at the end
of the retaining anyway)

In the process of doing this I reworked the hostport.Peer tests, but,
only removed a couple of tests from the Agent tests, I'll look into
fixing the agent tests later, but since there was no new functionality
in the agent I didn't see a need for rewriting the tests yet.