[GRPC][OnlyForBranch][Part 7] Add Caller and Encoding Headers to GRPC calls (#356)
* [GRPC][Part 4.5] Fix Lint after golint upgrade

Summary: In a recent commit to golint, there was a new rule added that
asserts that context.Context must be the first parameter in a function
definition.  This breaks some of GRPC's functionality since the handler
interface they require has ctx in the second position.  As a result, we
are fixing the cases where we have ctx in the second position, and we're
isolating the GRPC function we provide into it's own file and ignoring
it in our linting

* Fix the context that's being used

* s/handleImpl/handle/

* Add comment to makefile

* [GRPC][Part 5] Adding procedure routing on inbound calls

Summary: In our "barebones" version of the grpc inbound we did not have
the ability to call into any service/procedure other than "yarpc/yarpc".
This diff configures the gRPC services to reflect whatever we're using
in the YARPC Registry.  As a result we can perform routing to custom
service-procedures defined in yarpc

Test Plan: Updated the yclient and yserver to route to custom procedures
and tested locally

* Move serviceDesc generation into a separate function, added a comment to better explain what's going on

* Cleanup after inbound.Start() changes, and a little bit of other cleanup

* Fixup inbound slice creation, and passing fields by reference

* [GRPC][Part 6] Add First Crossdock test for GRPC

Summary: With the barebones functionality we currently have, we can only
support the raw usecase, but we have a crossdock test

* Add grpc to docker-compose.yaml

* [GRPC][Part 7] Add Caller and Encoding Headers to GRPC calls

Summary: Now that we have hardcoded raw functionality for yarpc over
gRPC, this diff add the ability to pass "Caller" and "Encoding" Headers
over the gRPC request.  In an ideal world, we would be able to set the
encoding directly in gRPC via a codec, but unfortunately it doesn't look
possible at this moment (maybe with a PR to gRPC).  The intermediary
state is a combination of the passthrough codec and adding this encoding
header so we can know what the encoding of the body is.

Test Plan: Added crossdock tests for json,thrift, and thriftgauntlet
proving we can handle the multiple encodings.  Also added a standalone
thrift example and ran the yclient/yserver to ensure the encodings &
callers were being pass through

* Remove thrift example, use custom errors for Header extraction

* Add tests for getContextMetadata and extractMetadataHeader