go:generate: Fix encoding/thrift corruption
Previously, running `go generate` in encoding/thrift would unconditionally
break the generated code because of vendored imports. This changes that
generate clause to use mockgen's source mode for that file. Source mode
doesn't work for everything but it does for this specific case.

This should allow putting `make generate` in our CI.