docker: Better caching and add `make docker-{build,test}` (#628)
This does a few things:

-   Adds a .dockerignore that matches .gitignore, specifically so that vendor
    is ignored.
-   Updates the Dockerfile with a few minor cleanups (1.7 -> 1.74, etc).
-   Updates the Dockerfile to do `glide install` in a cache-friendly way, and
    set WORKDIR.
-   Changes ENTRYPOINT to CMD to make it simpler to override the default
    command.
-   Adds `make docker-build` and `make docker-test`, which use Docker
    directly, but match the current name of the image outputted by docker-
    compose (because of port bindings, doing `docker-compose run go make test`
    will fail).

I like having a Dockerfile that allows me to run make test in a fresh
environment, to confirm that make test works out of the box for new
developers. With that said, it still doesn't, because vendor isn't checked in,
and glide install is not a dependency of make test, but at least the
Dockerfile somewhat documents that (and it is annoying to have glide install
run every time).