Delete yarpc.Headers
With the Req/ResMeta deletion, the user-facing API for sending request headers
is `yarpc.WithHeader`,

    client.Call(
        ..., yarpc.WithHeader("foo", "bar"), yarpc.WithHeader("baz", "qux"))

And the API for receiving request headers in handlers is
`yarpc.CallFromContext` with the `Header(k string) string` function.

    call := yarpc.CallFromContext(ctx)
    call.Header("foo")

    for _, k := range call.HeaderNames() {
        fmt.Println(call.Header(k))
    }

The only place we still use `yarpc.Headers` is to receive response headers for
outgoing calls:

    var h yarpc.Headers
    client.Call(..., yarpc.ResponseHeaders(&h))

This switches that to `map[string]string`, allowing us to drop the
`yarpc.Headers` type compeletely.