thriftrw: Add -gomock flag
This adds a `-gomock` flag to the ThriftRW plugin for YARPC. If supplied, we
will generate gomock-compatible mocks for the service clients.

For each service, we generate,

    package ${service}test

    type MockClient struct {  // implements ${service}client.Interface
        // ..
    }

    func NewMockClient(mockCtrl *gomock.Controller) *MockClient {
        // ...
    }

Note that the client mocks are named MockClient rather than MockInterface.
This is to leave room for adding fake/mock servers inside the same package
instead of creating a new package.