Add ServiceName field to transport.Outbounds
Summary: Currently, if you want to have multiple outbounds with the same
service name, or if you want to change the service name for an outbound
without changing the 'key' you use to grab the ClientConfig, you had to
jump through some hoops to do so.

This diff adds the ability to set a "ServiceName" on the
transport.Outbounds struct (should be called transport.OutboundSpec).
If you specify the "ServiceName" it will override the lookup key you
use when grabbing the ClientConfig.
```
outbounds['populous-test'] := transport.Outbounds{
      ServiceName: "populous",
      Unary: tchannel.NewOutbound(),
}

...
// Returns a ClientConfig with service 'populous'
dispatcher.ClientConfig('populous-test')
```