Implement arbitrary interface{} decoder
This implements an internal `decode` package that allows us to decode
any `interface{}` (more specifically, `map[interface{}]interface{}`)
into a more complex object.

This mostly relies on the `mapstructure` library for most of its
functionality with the exception of a custom `Decoder` interface which
types may implement for more complex decoding logic. This is very
similar to the `yaml.Unmarshaler` interface and how that behaves.

We will use this package for the config system to allow decoding
user-provided configurations (in `map[string]interface{}` form) into
something more complex which we need to build a `yarpc.Config`.