Drop presets and intermediate Builder types
This changes Configurator to return a yarpc.Config or yarpc.Dispatcher
directly instead of an intermediate config.Builder object.

Note that although we've lost the printable representation of the
configuration (config.Builder), it's fairly straightforward to track a
printable "configuration snapshot" as we're compiling the configuration.
This can be implemented in a follow-up PR if necessary.