Update HTTP transport to support MaxIdleConnsPerHost
Summary: I've been running into a large issue with the http transport
during benchmarks where we see transient errors like: "Can't assign
requested address" which means that we've exausted the ephemeral port
range.  After some investigation I found out it was because the go http
library sets a limit on the number of idle connections it keeps.

The parameter is MaxIdleConnsPerHost and The default is 2, which means
if we have more than 2 idle requests at any point we'll start churning
through ephemeral ports until we run out.