Create an internal io.Copy using sync.Pool byte slice buffers
Summary: By default, if you call io.Copy and you don't pass in a
WriterTo or a ReaderAt, the io.Copy function will allocate a byte buffer
to be an intermediary to copy the bytes into. This diff makes an
internal function that creates the intermediary buffer in a pool so that
it can be reused on subsequent requests.