Update lifecycle state management internals
Summary: During an intellectual excercise between @bombela and I we
went through the LifecycleOnce API to change it to use a single atomic
for state.  At the same time we noticed a bug in the functionality where
"Stop" would be called even if "Start" had never been called.

This diff changes the LifecycleOnce to use a singel atomic for state,
and updates the peer list implementations (which had bad assumptions
about what to do if the list hadn't been started yet) to block updates
to the list until the peer list had been started.  The blocks in the
update should block any watchers that attempt to write to the peerlist
so we should retain the order of updates.