Fix Data Race in LightMockPeer object
Summary: This fixes a race in a test version of Peer Objects.  The
status objects for peers does not use an atomic for pending count,
silly in retrospect now, but still teeeeechnically part of the api.

I would count fixing the peerStatus as a bug fix, but I'm curious what
other people think.  I don't think many people are using this API yet.