Fix Second race condition in lifecycle code
Summary: There is a race condition for access to the l.err in the
lifecycleOnce code.  Essentially, if we have the following occurence:
- Start
- Failed Stop
- Start

There is a potential race between the Failed Stop and the second Start.
Because we're sharing the err between start and stop (maybe that's an
issue possible to discuss that here).

I've put this as a separate diff from the singlestate refactor so we can
review it directly.