Remove limits from thrift decoder buffer pool
Summary: One of the first use cases that moved to this buffer did not
notice any significant change in their timings because their payloads
were over the limit we set.

Talked it over with @prashant and we agreed that we can remove the
buffer and add it back if it becomes an issue.