pally: auto-scrub invalid runes from label values
In order for vectors to be useful, we need to support auto-scrubbing invalid
runes from label values. For example, we might want to track a count of calls by
service; since service names may contain characters that aren't valid in Tally
(e.g., `!` or 🏦), we should automatically replace those characters with a
fallback.

This is very similar to the logic we use in Muttley.